backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
ortools==9.8.3296
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.13.1
